app/requirements.txt
flask
spotipy
yt-dlp
ffmpeg-python
mutagen
requests
python-dotenv
gunicorn
